Create perirority_encoder_generic
generic priority encoder module. It takes an input vector 'x' of size   [ n-1:0 ], an enable signal 'enb', and provides two outputs: 'v' and 'y'. 'v' is a single-bit output indicating whether any of the enabled inputs in 'x' are high, and 'y' is a binary-encoded output representing the highest enabled input.